+++ frontend/src/app/services/platform.service.ts
@@ -9,6 +9,7 @@ import {LighthouseSourceSearch} from '../models/lighthouse/lighthouse-source-sea
 import {environment} from '../../environments/environment';
 import {ResponseWrapper} from '../models/response-wrapper';
 import {map} from 'rxjs/operators';
+import {uuidV4} from '../../lib/utils/uuid';
 
 @Injectable({
   providedIn: 'root'
@@ -22,8 +23,8 @@ export class PlatformService {
       throw new Error("Invalid CCDA file")
     }
 
-    const endpointUrl = new URL(`https://api.platform.fastenhealth.com/v1/app/ccda-to-fhir`);
-    return this._httpClient.post<string>(endpointUrl.toString(), ccdaFile, { headers: {'Content-Type': 'application/cda+xml'} })
+    const endpointUrl = new URL(`https://api.platform.fastenhealth.com/v1/app/convert/ccda/to/fhir?patientId=${uuidV4()}`);
+    return this._httpClient.post<string>(endpointUrl.toString(), ccdaFile, { headers: {'Content-Type': 'application/xml', 'Accept':'application/json'} })
       .pipe(
         map((responseJson: string) => {
           return new File([JSON.stringify(responseJson)], ccdaFile.name + ".converted.json", {type: "application/json", lastModified: ccdaFile.lastModified || Date.now()})
